feat(3.4): extract all 54 routes into Flask blueprints in backend/api/
- backend/api/system.py   — 17 system/infra routes
- backend/api/gpu.py      — 2 GPU routes
- backend/api/costs.py    — 4 cost/power routes
- backend/api/experiments.py — 10 ML experiment routes
- backend/api/uptime_api.py  — 9 uptime/public-status routes
- backend/api/hosts_api.py   — 7 fleet/host routes
- backend/api/integrations.py — 4 notify/container routes
- backend/auth.py         — require_api_key helper (breaks circular import)

All 54 @app.route decorators removed from app.py; blueprints registered
via app.register_blueprint(). Lazy `import app as _app` inside each
handler preserves patch.object(app, ...) compatibility in tests.
Restored accidentally-removed _MLF_STATUS/_mlf helpers to app.py.

441 tests pass, 71 snapshots byte-identical.
