+++ app.py
@@ -80,6 +80,22 @@ REAL_FS      = {"ext4", "ext3", "xfs", "btrfs", "zfs", "vfat"}
 
 app = Flask(__name__, static_url_path="/static", static_folder="static")
 
+# Phase 3.4: register API blueprints (in original @app.route declaration order)
+from backend.api.system import bp as _system_bp
+from backend.api.gpu import bp as _gpu_bp
+from backend.api.costs import bp as _costs_bp
+from backend.api.experiments import bp as _experiments_bp
+from backend.api.uptime_api import bp as _uptime_api_bp
+from backend.api.hosts_api import bp as _hosts_api_bp
+from backend.api.integrations import bp as _integrations_bp
+app.register_blueprint(_system_bp)
+app.register_blueprint(_gpu_bp)
+app.register_blueprint(_costs_bp)
+app.register_blueprint(_experiments_bp)
+app.register_blueprint(_uptime_api_bp)
+app.register_blueprint(_hosts_api_bp)
+app.register_blueprint(_integrations_bp)
+
 # ── Prometheus gauges (defined once at module level) ──────────────────────────
 if _PROM_OK:
     _G = {
@@ -4005,7 +4021,6 @@ def _http_probe_once(target, timeout, method):
     return last_code
 
 
-
 def _tls_cert_days(host, port, timeout):
     """Return (days_remaining, expires_at_ts) or (None, None)."""
     import ssl, datetime, socket
@@ -4954,97 +4969,6 @@ def build_insights(total, services, mem_total, events, host):
 # ── API ──────────────────────────────────────────────────────────────────
 RANGES = {"1h": 3600, "6h": 21600, "24h": 86400, "7d": 604800, "30d": 2592000, "all": None}
 
-@app.route("/api/data")
-def api_data():
-    rng = request.args.get("range", "6h")
-    span = RANGES.get(rng, 21600); now = int(time.time())
-    with LOCK:
-        cur = DB.cursor()
-        since = (cur.execute("SELECT MIN(ts) FROM samples").fetchone()[0] or now) if span is None else now - span
-        bk = max(INTERVAL, round(max(1, now - since) / MAX_POINTS))
-        tot = cur.execute("SELECT (ts/?)*? b,AVG(util),AVG(mem_used),MAX(mem_used),AVG(power),AVG(temp),"
-                          "AVG(cpu),AVG(ram_used),AVG(ram_total),AVG(load1),AVG(ctemp) "
-                          "FROM samples WHERE ts>=? GROUP BY b ORDER BY b", (bk, bk, since)).fetchall()
-        labels = [int(r[0]) for r in tot]
-        idx = {b: i for i, b in enumerate(labels)}
-        total = {"util": [round(r[1] or 0) for r in tot], "mem": [round(r[2] or 0) for r in tot],
-                 "mempk": [round(r[3] or 0) for r in tot], "power": [round(r[4] or 0) for r in tot],
-                 "temp": [round(r[5] or 0) for r in tot], "cpu": [round(r[6] or 0) for r in tot],
-                 "ram_used": [round(r[7] or 0) for r in tot], "ram_total": [round(r[8] or 0) for r in tot],
-                 "load1": [round(r[9] or 0, 2) for r in tot], "ctemp": [round(r[10] or 0) for r in tot]}
-        services = {}
-        for b, svc, mem in cur.execute("SELECT (ts/?)*? b,service,AVG(mem) FROM proc WHERE ts>=? GROUP BY b,service",
-                                       (bk, bk, since)).fetchall():
-            i = idx.get(int(b))
-            if i is not None:
-                services.setdefault(svc, [0] * len(labels))[i] = round(mem or 0)
-        other = [max(0, total["mem"][i] - sum(s[i] for s in services.values())) for i in range(len(labels))]
-        # Per-device disk-I/O trend, same bucketing/labels as everything else on this
-        # endpoint — feeds the Disk I/O tab's per-device sparklines. disk_io_samples
-        # rides its own ~45s cadence (sparser than `samples`), so buckets it doesn't
-        # cover keep their 0-fill default (matches the `services` fill above).
-        disk_io = {}
-        for b, dev, r, w, u in cur.execute(
-                "SELECT (ts/?)*? b,device,AVG(read_mb_s),AVG(write_mb_s),AVG(util_pct) "
-                "FROM disk_io_samples WHERE ts>=? GROUP BY b,device", (bk, bk, since)).fetchall():
-            i = idx.get(int(b))
-            if i is None:
-                continue
-            d = disk_io.setdefault(dev, {"read_mb_s": [0] * len(labels),
-                                         "write_mb_s": [0] * len(labels),
-                                         "util_pct": [0] * len(labels)})
-            d["read_mb_s"][i]  = round(r or 0, 1)
-            d["write_mb_s"][i] = round(w or 0, 1)
-            d["util_pct"][i]   = round(u or 0, 1)
-        ticks = cur.execute("SELECT COUNT(*) FROM samples WHERE ts>=?", (since,)).fetchone()[0] or 1
-        summary = sorted(({"service": s, "peak": round(pk), "avg": round(av), "present": round(100 * cnt / ticks)}
-                          for s, pk, av, cnt in cur.execute(
-                              "SELECT service,MAX(mem),AVG(mem),COUNT(DISTINCT ts) FROM proc WHERE ts>=? GROUP BY service",
-                              (since,)).fetchall()), key=lambda x: -x["peak"])
-        model_summary = sorted(({"service": s, "model": m, "peak": round(pk or 0), "avg": round(av or 0)}
-                                for s, m, pk, av in cur.execute(
-                                    "SELECT service,model,MAX(vram),AVG(vram) FROM models WHERE ts>=? AND vram IS NOT NULL "
-                                    "GROUP BY service,model", (since,)).fetchall()), key=lambda x: -x["peak"])
-        # Caller attribution: connection-seconds per (caller → server) over the range.
-        # Each sample of `conns` open connections represents INTERVAL seconds of traffic.
-        callers = sorted(({"caller": c, "server": s, "seconds": int((tot or 0) * INTERVAL), "samples": n}
-                          for c, s, tot, n in cur.execute(
-                              "SELECT caller,server,SUM(conns),COUNT(DISTINCT ts) FROM edges WHERE ts>=? "
-                              "GROUP BY caller,server", (since,)).fetchall()), key=lambda x: -x["seconds"])
-        evs = [{"ts": t, "service": s, "kind": k, "detail": d}
-               for t, s, k, d in cur.execute("SELECT ts,service,kind,detail FROM events WHERE ts>=? ORDER BY ts",
-                                              (since,)).fetchall()]
-        oom_evs = [e for e in evs if e["kind"] == "oom"]
-        for e in oom_evs:
-            row = cur.execute("SELECT service,mem FROM proc WHERE ts<=? AND service!=? ORDER BY ts DESC,mem DESC LIMIT 1",
-                              (e["ts"] + INTERVAL, e["service"])).fetchone()
-            if row:
-                e["blame"] = (f"{e['service']} lost to {row[0]} (holding {round(row[1])} MB) at "
-                              f"{time.strftime('%Y-%m-%d %H:%M', time.localtime(e['ts']))}.")
-        mem_total = LATEST["mem_total"] or 24576
-        peak = max(total["mempk"]) if total["mempk"] else 0
-        insights = build_insights(total, services, mem_total, oom_evs, LATEST["host"])
-        diskio_evs = [e for e in evs if e["kind"] == "diskio_spike"]
-        if diskio_evs:
-            latest_by_dev = {}
-            for e in diskio_evs:
-                latest_by_dev[e["service"]] = e   # `service` column holds the device name here
-            for dev, e in latest_by_dev.items():
-                insights.append({"level": "warning", "title": f"Disk I/O spike on {dev}",
-                                 "detail": e["detail"]})
-    # Uptime rows ride the same Insight Feed (computed outside LOCK — uptime_overview
-    # takes it itself). DOWN/slow endpoints surface on the cockpit with no new tile.
-    try:
-        insights = insights + uptime_insights()
-        up_summary = uptime_summary()
-    except Exception as e:
-        print("uptime overview error:", e, flush=True)
-        up_summary = {"total": 0, "up": 0, "down": 0, "unknown": 0, "worst_down": None}
-    return jsonify({"version": VERSION, "range": rng, "bucket_sec": bk, "labels": labels, "total": total,
-                    "services": services, "other": other, "summary": summary, "model_summary": model_summary,
-                    "callers": callers, "events": oom_evs, "insights": insights, "pressure_free_mb": PRESSURE_MB,
-                    "uptime_summary": up_summary, "disk_io": disk_io,
-                    "mem_total": mem_total, "peak_mem": peak, "now": LATEST})
 
 def _hhmm_to_min(s, default):
     """'HH:MM' -> minutes since midnight [0,1440). Falls back to `default` on junk."""
@@ -5074,105 +4998,7 @@ def _make_is_night(night_start, night_end):
             return mins >= ns or mins < ne
     return is_night
 
-@app.route("/api/cost")
-def api_cost():
-    """Power → kWh → money (#25), now tariff-aware. Integrates the GPU `power`
-    samples we already collect; each sample stands for INTERVAL seconds, so
-    energy(kWh) = sum(power_W) * INTERVAL / 3_600_000.
 
-    Single mode (default): cost = energy * kwh_price — byte-for-byte the original
-    behaviour. Dual mode: each sample is billed at the night price inside the
-    (possibly midnight-wrapping) night window and the day price otherwise, split
-    per window. A blank night price silently degrades to single, so a user who
-    doesn't know their rates keeps the simple average. The card stays hidden until
-    a day price is set (`enabled`)."""
-    s = get_settings()
-    def fnum(key):
-        v = (s.get(key) or "").strip()
-        if v == "":
-            return None
-        try:
-            return float(v)
-        except ValueError:
-            return None
-
-    day_price   = fnum("kwh_price") or 0.0
-    night_price = fnum("kwh_price_night")
-    mode = "dual" if (s.get("tariff_mode") == "dual" and night_price is not None
-                      and day_price > 0) else "single"
-    currency = s.get("currency") or "$"
-    is_night = _make_is_night(s.get("night_start", "22:00"), s.get("night_end", "06:00"))
-
-    rng = request.args.get("range", "7d")
-    span = RANGES.get(rng, 604800)
-    now = int(time.time())
-    kwh_per_wsample = INTERVAL / 3_600_000.0   # one power sample -> kWh
-
-    with LOCK:
-        cur = DB.cursor()
-        def avg_w(since):
-            return round(cur.execute("SELECT AVG(power) FROM samples WHERE ts>=?", (since,)).fetchone()[0] or 0)
-        def total_kwh(since):
-            tot = cur.execute("SELECT SUM(power*cnt) FROM samples_1h WHERE ts>=?", (since,)).fetchone()[0] or 0
-            return tot * kwh_per_wsample
-        def split_kwh(since):
-            """One pass over (ts,power,cnt) >= since -> (day_kwh, night_kwh)."""
-            day_w = night_w = 0.0
-            for ts, p, c in cur.execute("SELECT ts,power,cnt FROM samples_1h WHERE ts>=? AND power IS NOT NULL", (since,)):
-                if is_night(ts):
-                    night_w += (p or 0) * (c or 1)
-                else:
-                    day_w += (p or 0) * (c or 1)
-            return day_w * kwh_per_wsample, night_w * kwh_per_wsample
-
-        lt = time.localtime(now)
-        midnight = int(time.mktime((lt.tm_year, lt.tm_mon, lt.tm_mday, 0, 0, 0, 0, 0, -1)))
-        wins = {"today": midnight, "d7": now - 604800, "d30": now - 2592000}
-        kwh, cost, split = {}, {}, {}
-        for w, since in wins.items():
-            if mode == "dual":
-                dk, nk = split_kwh(since)
-            else:
-                dk, nk = total_kwh(since), 0.0       # single: one SUM, no per-row loop
-            dc, nc = dk * day_price, nk * (night_price or 0.0)
-            kwh[w]  = round(dk + nk, 3)
-            cost[w] = round(dc + nc, 2)
-            split[w] = {"day_kwh": round(dk, 3), "night_kwh": round(nk, 3),
-                        "day_cost": round(dc, 2), "night_cost": round(nc, 2)}
-
-        # Cumulative-cost series across the selected range (mirrors api_data buckets).
-        since = (cur.execute("SELECT MIN(ts) FROM samples_1h").fetchone()[0] or now) if span is None else now - span
-        bk = max(INTERVAL, round(max(1, now - since) / MAX_POINTS))
-        labels, cost_cum, running = [], [], 0.0
-        if mode == "dual":                            # stream + classify per bucket (one pass)
-            acc = {}
-            for ts, p, c in cur.execute("SELECT ts,power,cnt FROM samples_1h WHERE ts>=? AND power IS NOT NULL ORDER BY ts", (since,)):
-                b = (ts // bk) * bk
-                price = night_price if is_night(ts) else day_price
-                acc[b] = acc.get(b, 0.0) + (p or 0) * (c or 1) * kwh_per_wsample * price
-            for b in sorted(acc):
-                running += acc[b]
-                labels.append(int(b)); cost_cum.append(round(running, 4))
-        else:                                         # single: cheap SQL-bucketed path
-            rows = cur.execute("SELECT (ts/?)*? b, SUM(power*cnt) FROM samples_1h WHERE ts>=? GROUP BY b ORDER BY b",
-                               (bk, bk, since)).fetchall()
-            for b, p in rows:
-                running += (p or 0) * kwh_per_wsample * day_price
-                labels.append(int(b)); cost_cum.append(round(running, 4))
-
-    return jsonify({
-        "enabled": day_price > 0, "kwh_price": day_price, "currency": currency,
-        "range": rng, "bucket_sec": bk,
-        "current_w": round(LATEST.get("power") or 0),
-        "avg_24h_w": avg_w(now - 86400), "avg_7d_w": avg_w(now - 604800),
-        "kwh": kwh, "cost": cost, "split": split,
-        "tariff": {"mode": mode, "price_day": day_price, "price_night": night_price,
-                   "night_start": s.get("night_start", "22:00"),
-                   "night_end": s.get("night_end", "06:00")},
-        "series": {"labels": labels, "cost_cum": cost_cum},
-    })
-
-# ── Costs page: per-machine → per-component → per-process (with drilldown) ─────
 _TOTAL_W_EXPR = "COALESCE(power,0)+COALESCE(cpu_power,0)+COALESCE(dram_power,0)"
 
 def _cost_ctx():
@@ -5196,307 +5022,7 @@ def _cost_ctx():
 def _price_at(ctx, ts):
     return ctx["night"] if (ctx["mode"] == "dual" and ctx["is_night"](ts)) else ctx["day"]
 
-@app.route("/api/costs")
-def api_costs():
-    """Richer power+cost view for the Costs page: per-machine totals, a stacked
-    component breakdown (GPU measured, CPU/DRAM measured via RAPL, optional operator
-    'other' baseline) and a ranked per-process/service/model breakdown — all over a
-    selectable range and tariff-aware. /api/cost (GPU-only) is left untouched."""
-    ctx = _cost_ctx()
-    cur = ctx["currency"]
-    rng = request.args.get("range", "7d")
-    span = RANGES.get(rng, 604800)
-    now = int(time.time())
-    kwh_per = INTERVAL / 3_600_000.0
-    with LOCK:
-        c = DB.cursor()
-        since = (c.execute("SELECT MIN(ts) FROM samples_1h").fetchone()[0] or now) if span is None else now - span
-        bk = max(INTERVAL, round(max(1, now - since) / MAX_POINTS))
-        comp = c.execute(f"SELECT (ts/?)*? b, AVG(power), AVG(cpu_power), AVG(dram_power) "
-                         f"FROM samples_1h WHERE ts>=? GROUP BY b ORDER BY b", (bk, bk, since)).fetchall()
-        # component energy + cost over the range (tariff-aware, one streaming pass)
-        comp_kwh = {"gpu": 0.0, "cpu": 0.0, "dram": 0.0}
-        cost_range = 0.0
-        nticks = 0
-        for ts, p, cp, dp, cnt_ in c.execute("SELECT ts,power,cpu_power,dram_power,cnt FROM samples_1h WHERE ts>=?", (since,)):
-            nticks += cnt_ or 1
-            price = _price_at(ctx, ts)
-            n = cnt_ or 1
-            tot = ((p or 0) + (cp or 0) + (dp or 0)) * n
-            comp_kwh["gpu"] += (p or 0) * n * kwh_per
-            comp_kwh["cpu"] += (cp or 0) * n * kwh_per
-            comp_kwh["dram"] += (dp or 0) * n * kwh_per
-            cost_range += tot * kwh_per * price
-        # today/d7/d30 total-cost windows (machine total watts, tariff-aware)
-        def win_cost(start):
-            tot = 0.0
-            for ts, w, cnt_ in c.execute(f"SELECT ts, {_TOTAL_W_EXPR} w, cnt FROM samples_1h WHERE ts>=?", (start,)):
-                tot += (w or 0) * (cnt_ or 1) * kwh_per * _price_at(ctx, ts)
-            return round(tot, 2)
-        lt = time.localtime(now)
-        midnight = int(time.mktime((lt.tm_year, lt.tm_mon, lt.tm_mday, 0, 0, 0, 0, 0, -1)))
-        cost_win = {"today": win_cost(midnight), "d7": win_cost(now - 604800), "d30": win_cost(now - 2592000)}
-        # ranked per-entity breakdown from power_proc (tariff-aware day/night split)
-        acc = {}
-        for ts, kind, name, watts in c.execute("SELECT ts,kind,name,watts FROM power_proc WHERE ts>=?", (since,)):
-            a = acc.setdefault((kind, name), [0.0, 0.0])
-            if ctx["mode"] == "dual" and ctx["is_night"](ts):
-                a[1] += watts
-            else:
-                a[0] += watts
-    hours = max(1e-9, (now - since) / 3600.0)
-    idle_w = ctx["idle_w"]
-    labels = [int(r[0]) for r in comp]
-    series = {"labels": labels,
-              "gpu":  [round(r[1] or 0) for r in comp],
-              "cpu":  [round(r[2] or 0) if r[2] is not None else 0 for r in comp],
-              "dram": [round(r[3] or 0) if r[3] is not None else 0 for r in comp]}
-    have_cpu = any(r[2] is not None for r in comp)
-    have_dram = any(r[3] is not None for r in comp)
-    if not have_cpu: series.pop("cpu")
-    if not have_dram: series.pop("dram")
-    if idle_w:
-        series["other"] = [round(idle_w)] * len(labels)
-    breakdown = []
-    for (kind, name), (dayw, nightw) in acc.items():
-        energy = (dayw + nightw) * kwh_per
-        cost = (dayw * ctx["day"] + nightw * (ctx["night"] if ctx["night"] is not None else ctx["day"])) * kwh_per
-        breakdown.append({"kind": kind, "name": name, "energy_kwh": round(energy, 4),
-                          "cost": round(cost, 4), "avg_w": round((dayw + nightw) / max(1, nticks))})
-    breakdown.sort(key=lambda x: -x["energy_kwh"])
-    now_gpu = round(LATEST.get("power") or 0)
-    now_cpu = round(LATEST.get("cpu_power") or 0) if LATEST.get("cpu_power") is not None else None
-    now_dram = round(LATEST.get("dram_power") or 0) if LATEST.get("dram_power") is not None else None
-    now_total = now_gpu + (now_cpu or 0) + (now_dram or 0) + (round(idle_w) if idle_w else 0)
-    measured = ["gpu"] + (["cpu"] if have_cpu else []) + (["dram"] if have_dram else [])
-    machine = {"name": "local",
-               "now_w": {"gpu": now_gpu, "cpu": now_cpu, "dram": now_dram, "total": now_total},
-               "energy_kwh": {k: round(v, 3) for k, v in comp_kwh.items() if (k != "dram" or have_dram)},
-               "cost": cost_win, "cost_range": round(cost_range, 2),
-               "measured": measured, "estimated": (["other"] if idle_w else [])}
-    machine["energy_kwh"]["total"] = round(sum(machine["energy_kwh"][k] for k in machine["energy_kwh"] if k != "total"), 3)
-    return jsonify({
-        "enabled": ctx["day"] > 0, "range": rng, "bucket_sec": bk, "currency": cur,
-        "rapl_available": have_cpu,
-        "tariff": {"mode": ctx["mode"], "price_day": ctx["day"], "price_night": ctx["night"],
-                   "night_start": ctx["night_start"], "night_end": ctx["night_end"]},
-        "machines": [machine], "components": series, "breakdown": breakdown[:40],
-    })
 
-@app.route("/api/cost/heatmap")
-def api_cost_heatmap():
-    """Busy-vs-quiet rhythm of the lab as a 7×24 grid (local day-of-week × hour).
-
-    Each historical `samples` row is the machine's total draw (GPU+CPU+DRAM) for
-    one INTERVAL tick. We bucket every tick by its LOCAL weekday/hour and average
-    the watts in each cell, then derive a cost-rate (€/h) for that cell at the
-    tariff's price for that hour band — reusing the same `_cost_ctx`/`_price_at`
-    machinery as the Costs page, so the €/kWh math never diverges. Sparse cells
-    carry their own sample count so the UI can be honest about coverage.
-
-    Pure-Python aggregation, read outside any held lock, always 200. When cost is
-    disabled (no tariff) we still return the power grid and `enabled:false` so the
-    card can render watts and prompt for a price.
-    """
-    ctx = _cost_ctx()
-    cur = ctx["currency"]
-    # window: last N days, sane default 30, capped at a year so a huge DB can't stall
-    try:
-        days = int(request.args.get("days", "30"))
-    except (TypeError, ValueError):
-        days = 30
-    days = max(1, min(days, 365))
-    now = int(time.time())
-    since = now - days * 86400
-
-    # 7×24 accumulators: summed watts and tick count per local day/hour cell
-    sum_w = [[0.0] * 24 for _ in range(7)]
-    cnt   = [[0]   * 24 for _ in range(7)]
-    span_min = span_max = None
-    try:
-        with LOCK:
-            c = DB.cursor()
-            rows = c.execute(
-                f"SELECT ts, {_TOTAL_W_EXPR} w, cnt FROM samples_1h WHERE ts>=? ORDER BY ts",
-                (since,)).fetchall()
-        # aggregate OUTSIDE the lock — pure Python, no DB calls below
-        for ts, w, row_cnt in rows:
-            lt = time.localtime(ts)
-            # Python weekday(): Mon=0..Sun=6 — matches our locale day labels
-            d = lt.tm_wday
-            h = lt.tm_hour
-            sum_w[d][h] += (w or 0) * (row_cnt or 1)
-            cnt[d][h] += row_cnt or 1
-            if span_min is None or ts < span_min:
-                span_min = ts
-            if span_max is None or ts > span_max:
-                span_max = ts
-    except Exception:
-        rows = []
-
-    # build the grids; price each cell at the tariff for a representative ts in it
-    rep_ts = span_max or now
-    rep_lt = time.localtime(rep_ts)
-    # anchor to local midnight of the most recent observed day so is_night() lands
-    # in the right band per (day,hour); the date component is irrelevant to the band
-    anchor = int(time.mktime((rep_lt.tm_year, rep_lt.tm_mon, rep_lt.tm_mday,
-                              0, 0, 0, 0, 0, -1)))
-    avg_w  = [[None] * 24 for _ in range(7)]
-    cost_h = [[None] * 24 for _ in range(7)]   # cost per HOUR at this cell's mean draw
-    max_w = max_cost = 0.0
-    busiest = quietest = None                  # by avg watts
-    total_ticks = 0
-    for d in range(7):
-        for h in range(24):
-            n = cnt[d][h]
-            total_ticks += n
-            if n == 0:
-                continue
-            aw = sum_w[d][h] / n
-            avg_w[d][h] = round(aw)
-            # price for this hour band: a ts at hour h on the anchor day
-            cell_ts = anchor + h * 3600
-            price = _price_at(ctx, cell_ts)
-            ch = aw / 1000.0 * price           # W -> kW * €/kWh = €/h
-            cost_h[d][h] = round(ch, 4)
-            max_w = max(max_w, aw)
-            max_cost = max(max_cost, ch)
-            if busiest is None or aw > busiest["avg_w"]:
-                busiest = {"day": d, "hour": h, "avg_w": round(aw),
-                           "cost_h": round(ch, 4), "samples": n}
-            if quietest is None or aw < quietest["avg_w"]:
-                quietest = {"day": d, "hour": h, "avg_w": round(aw),
-                            "cost_h": round(ch, 4), "samples": n}
-
-    # busy vs quiet bands: top/bottom quartile of populated cells by avg watts
-    populated = [(avg_w[d][h], cost_h[d][h])
-                 for d in range(7) for h in range(24) if avg_w[d][h] is not None]
-    bands = None
-    if len(populated) >= 4:
-        ordered = sorted(populated, key=lambda x: x[0])
-        q = max(1, len(ordered) // 4)
-        quiet_band = ordered[:q]
-        busy_band = ordered[-q:]
-        def band_stats(b):
-            return {"avg_w": round(sum(x[0] for x in b) / len(b)),
-                    "avg_cost_h": round(sum((x[1] or 0) for x in b) / len(b), 4),
-                    "cells": len(b)}
-        bands = {"busy": band_stats(busy_band), "quiet": band_stats(quiet_band)}
-
-    # per-day rollups (busiest / quietest day by mean watts across populated hours)
-    day_avg = [None] * 7
-    for d in range(7):
-        vals = [avg_w[d][h] for h in range(24) if avg_w[d][h] is not None]
-        if vals:
-            day_avg[d] = round(sum(vals) / len(vals))
-
-    coverage = round(total_ticks / max(1, days * 24 * 3600 / INTERVAL), 4)
-    # "ready" once we have at least a day's worth of ticks spread across cells
-    populated_cells = len(populated)
-    ready = total_ticks >= (86400 / INTERVAL) and populated_cells >= 6
-
-    return jsonify({
-        "ok": True,
-        "enabled": ctx["day"] > 0,
-        "currency": cur,
-        "days": days,
-        "interval_sec": INTERVAL,
-        "ready": ready,
-        "rows": 7, "cols": 24,
-        "avg_w": avg_w,
-        "cost_h": cost_h,
-        "samples": cnt,
-        "day_avg_w": day_avg,
-        "max_w": round(max_w),
-        "max_cost_h": round(max_cost, 4),
-        "busiest": busiest,
-        "quietest": quietest,
-        "bands": bands,
-        "total_ticks": total_ticks,
-        "populated_cells": populated_cells,
-        "coverage": coverage,
-        "span": {"min": span_min, "max": span_max},
-        "tariff": {"mode": ctx["mode"], "price_day": ctx["day"], "price_night": ctx["night"],
-                   "night_start": ctx["night_start"], "night_end": ctx["night_end"]},
-    })
-
-@app.route("/api/costs/entity")
-def api_costs_entity():
-    """Per-entity drilldown: a power + cumulative-cost time-series for one
-    process/service/model over the range, plus what resources it used."""
-    ctx = _cost_ctx()
-    name = request.args.get("name", "")
-    kind = request.args.get("kind", "")
-    rng = request.args.get("range", "7d")
-    span = RANGES.get(rng, 604800)
-    now = int(time.time())
-    kwh_per = INTERVAL / 3_600_000.0
-    with LOCK:
-        c = DB.cursor()
-        since = (c.execute("SELECT MIN(ts) FROM power_proc").fetchone()[0] or now) if span is None else now - span
-        bk = max(INTERVAL, round(max(1, now - since) / MAX_POINTS))
-        q = "SELECT (ts/?)*? b, AVG(watts), MAX(watts) FROM power_proc WHERE name=? AND ts>=?"
-        args = [bk, bk, name, since]
-        if kind:
-            q += " AND kind=?"; args.append(kind)
-        q += " GROUP BY b ORDER BY b"
-        rows = c.execute(q, args).fetchall()
-        # cumulative tariff-aware cost needs per-bucket price; classify by bucket start ts
-        vram_peak = None
-        if kind != "cpu":
-            vram_peak = c.execute("SELECT MAX(mem) FROM proc WHERE service=? AND ts>=?", (name, since)).fetchone()[0]
-    labels, watts, cost_cum, running, energy = [], [], [], 0.0, 0.0
-    peak = 0.0
-    for b, avgw, maxw in rows:
-        labels.append(int(b)); watts.append(round(avgw or 0))
-        peak = max(peak, maxw or 0)
-        e = (avgw or 0) * kwh_per * (bk / INTERVAL)   # energy this bucket (avg W over bk seconds)
-        energy += e
-        running += e * _price_at(ctx, int(b))
-        cost_cum.append(round(running, 4))
-    return jsonify({
-        "name": name, "kind": kind, "range": rng, "bucket_sec": bk, "currency": ctx["currency"],
-        "energy_kwh": round(energy, 4), "cost": round(running, 2),
-        "avg_w": round(sum(watts) / len(watts)) if watts else 0, "peak_w": round(peak),
-        "series": {"labels": labels, "watts": watts, "cost_cum": cost_cum},
-        "resources": {"gpu_vram_peak_mb": round(vram_peak) if vram_peak else None},
-    })
-
-@app.route("/api/sessions")
-def api_sessions():
-    """GPU activity sessions over the range — contiguous GPU-busy periods rebuilt
-    from the power/util history. Plus the live training processes detected on the
-    hub right now (LATEST['training']). Powers the Experiments tab."""
-    s = get_settings()
-    try:
-        price = float(s.get("kwh_price") or 0)
-    except (TypeError, ValueError):
-        price = 0.0
-    rng = request.args.get("range", "7d")
-    span = RANGES.get(rng, 604800)
-    now = int(time.time())
-    with LOCK:
-        cur = DB.cursor()
-        since = (cur.execute("SELECT MIN(ts) FROM samples").fetchone()[0] or now) if span is None else now - span
-        rows = cur.execute("SELECT ts,util,power,mem_used FROM samples WHERE ts>=? ORDER BY ts", (since,)).fetchall()
-    sessions = _gpu_sessions(rows, INTERVAL, price=price)[:50]
-    tot_energy = round(sum(x["energy_kwh"] for x in sessions), 3)
-    return jsonify({
-        "range": rng, "currency": s.get("currency") or "$",
-        "price": price, "kwh_price": price,
-        "active_pct": _ACTIVE_UTIL,
-        "sessions": sessions,
-        "totals": {"count": len(sessions), "energy_kwh": tot_energy,
-                   "cost": round(tot_energy * price, 2),
-                   "active_hours": round(sum(x["duration"] for x in sessions) / 3600.0, 1)},
-        "training": LATEST.get("training") or [],
-        "devtools": LATEST.get("devtools") or [],
-    })
-
-# ── Integrations: experiment/run tracking API (push/pull) + MLflow sync ────────
-# Pivot from /proc auto-detection to a key-authenticated ingest API a notebook can
-# push to (Jupyter/Colab/Kaggle via homelab_run.py), pulled back with the run's real
-# GPU energy/cost attached by overlapping its [start,end] window with `samples`.
 MAX_RUN_FIELD, MAX_RUN_JSON, MAX_METRICS_REQ = 4096, 64 * 1024, 1000
 RUN_SOURCES = {"jupyter", "colab", "kaggle", "mlflow", "api", "cli"}
 RUN_STATUS  = {"running", "finished", "failed", "killed"}
@@ -5523,40 +5049,8 @@ def _gen_api_key(name="default", expires_in_days=None):
     """Back-compat shim (used by tests): mint a key and return the plaintext."""
     return _create_api_key(name, expires_in_days)[1]
 
-def _key_lookup(presented):
-    """Return the id of a live (non-expired) key matching `presented`, else None,
-    and stamp its last_used_at. Stored value is a hash, so the lookup never exposes
-    a usable secret. Fail-closed: no keys => all ingest rejected."""
-    if not presented:
-        return None
-    now = int(time.time())
-    with LOCK:
-        row = DB.execute("SELECT id, expires_at FROM api_keys WHERE key_hash=?",
-                         (_hash_key(presented),)).fetchone()
-        if not row:
-            return None
-        kid, exp = row
-        if exp and exp < now:
-            return None
-        DB.execute("UPDATE api_keys SET last_used_at=? WHERE id=?", (now, kid))
-        DB.commit()
-    return kid
-
-def _presented_key():
-    auth = request.headers.get("Authorization", "")
-    if auth[:7].lower() == "bearer ":
-        return auth[7:].strip()
-    return request.headers.get("X-API-Key", "").strip()
-
-def require_api_key(fn):
-    @wraps(fn)
-    def wrapper(*a, **kw):
-        kid = _key_lookup(_presented_key())
-        if not kid:
-            return jsonify({"ok": False, "error": "missing, invalid, or expired API key"}), 401
-        g.api_key_id = kid
-        return fn(*a, **kw)
-    return wrapper
+# Phase 3.4: moved to backend/auth — re-exported for backward compat
+from backend.auth import _key_lookup, _presented_key, require_api_key
 
 def _clip(v, n):
     return ("" if v is None else str(v))[:n]
@@ -5774,234 +5268,6 @@ def _merge_registry(ollama_models, catalog):
     return out
 
 
-@app.route("/api/models")
-def api_models():
-    """The Model Registry: the full inventory of models available on this host —
-    ollama's on-disk catalogue (GET /api/tags, size/quant/param detail) merged with
-    every other recognised AI server's model list (#219: vLLM, llama.cpp, LM Studio,
-    ComfyUI, InvokeAI, …), cross-referenced with what's loaded right now so the UI
-    can flag resident models + their live VRAM, grouped by provider.
-
-    Always 200, graceful-degrade, never 500, no secret leak (we echo a `reachable`
-    bool, never the URL/creds). Ollama half cached ~45s so a busy tab can't hammer
-    it; the rest rides the existing sampler's cached model_catalog (no extra I/O).
-    /api/tags is polled outside any held LOCK."""
-    ollama_models, reachable = _model_registry()
-    models = _merge_registry(ollama_models, LATEST.get("model_catalog"))
-    return jsonify({
-        "enabled": COPILOT_ENABLED,
-        "ollama_reachable": reachable,
-        "models": models,
-        "totals": _registry_totals(models),
-        "providers": sorted({m["provider"] for m in models}),
-    })
-
-
-@app.route("/api/integration/keys", methods=["GET", "POST"])
-def api_keys_route():
-    """GET -> {keys:[{id,name,prefix,created_at,expires_at,last_used_at,expired,runs}]}
-    (never the secret). POST {name, expires_in_days?} -> {id, key} (key revealed once)."""
-    if request.method == "POST":
-        body = request.get_json(silent=True) or {}
-        days = body.get("expires_in_days")
-        try:
-            days = int(days) if days not in (None, "", 0, "0") else None
-        except (TypeError, ValueError):
-            days = None
-        if days is not None and days <= 0:
-            days = None
-        kid, key = _create_api_key(_clip(body.get("name") or "key", 128), days)
-        return jsonify({"ok": True, "id": kid, "key": key})
-    now = int(time.time())
-    with LOCK:
-        rows = DB.execute("SELECT id,name,prefix,created_at,expires_at,last_used_at "
-                          "FROM api_keys ORDER BY created_at DESC").fetchall()
-        counts = dict(DB.execute("SELECT key_id, COUNT(*) FROM runs WHERE key_id IS NOT NULL "
-                                 "GROUP BY key_id").fetchall())
-    keys = [{"id": kid, "name": name, "prefix": prefix, "created_at": created,
-             "expires_at": exp, "last_used_at": used, "expired": bool(exp and exp < now),
-             "runs": counts.get(kid, 0)}
-            for (kid, name, prefix, created, exp, used) in rows]
-    return jsonify({"keys": keys, "has_key": bool(keys)})
-
-@app.route("/api/integration/keys/<kid>", methods=["DELETE"])
-def api_keys_delete(kid):
-    """Revoke (remove) a key. Runs it pushed are kept; they just lose live attribution."""
-    with LOCK:
-        cur = DB.execute("DELETE FROM api_keys WHERE id=?", (kid,))
-        DB.commit()
-    return (jsonify({"ok": True}) if cur.rowcount
-            else (jsonify({"ok": False, "error": "unknown key"}), 404))
-
-@app.route("/api/runs", methods=["POST"])
-@require_api_key
-def api_runs_create():
-    body = request.get_json(silent=True) or {}
-    source = (body.get("source") or "api").lower()
-    if source not in RUN_SOURCES:
-        source = "api"
-    now = int(time.time())
-    rid = (body.get("id") or uuid.uuid4().hex)[:64]
-    try:
-        params = _json_field(body.get("params"), MAX_RUN_JSON)
-        tags = _json_field(body.get("tags"), MAX_RUN_JSON)
-    except ValueError as e:
-        return jsonify({"ok": False, "error": str(e)}), 413
-    with LOCK:
-        DB.execute("INSERT INTO runs(id,name,source,status,started_at,ended_at,host,params,tags,notes,"
-                   "heartbeat_at,ext_id,created_at,key_id) VALUES(?,?,?,?,?,?,?,?,?,?,?,?,?,?) "
-                   "ON CONFLICT(id) DO NOTHING",
-                   (rid, _clip(body.get("name") or "run", MAX_RUN_FIELD), source, "running",
-                    int(body.get("started_at") or now), None, _clip(body.get("host"), 256),
-                    params, tags, _clip(body.get("notes"), MAX_RUN_FIELD), now, None, now,
-                    getattr(g, "api_key_id", None)))
-        DB.commit()
-    return jsonify({"ok": True, "id": rid})
-
-@app.route("/api/runs/<rid>", methods=["PATCH"])
-@require_api_key
-def api_runs_update(rid):
-    body = request.get_json(silent=True) or {}
-    sets, args = ["heartbeat_at=?"], [int(time.time())]
-    if body.get("status") in RUN_STATUS:
-        sets.append("status=?"); args.append(body["status"])
-    if body.get("ended_at"):
-        sets.append("ended_at=?"); args.append(int(body["ended_at"]))
-    for f, n in (("name", MAX_RUN_FIELD), ("notes", MAX_RUN_FIELD)):
-        if f in body:
-            sets.append(f"{f}=?"); args.append(_clip(body[f], n))
-    try:
-        for f in ("params", "tags"):
-            if f in body:
-                sets.append(f"{f}=?"); args.append(_json_field(body[f], MAX_RUN_JSON))
-    except ValueError as e:
-        return jsonify({"ok": False, "error": str(e)}), 413
-    args.append(rid)
-    with LOCK:
-        cur = DB.execute(f"UPDATE runs SET {','.join(sets)} WHERE id=?", args)
-        DB.commit()
-    return (jsonify({"ok": True}) if cur.rowcount else (jsonify({"ok": False, "error": "unknown run"}), 404))
-
-@app.route("/api/runs/<rid>/metrics", methods=["POST"])
-@require_api_key
-def api_runs_metrics(rid):
-    body = request.get_json(silent=True) or {}
-    pts = body.get("metrics")
-    if pts is None and "key" in body:
-        pts = [body]
-    pts = pts or []
-    if len(pts) > MAX_METRICS_REQ:
-        return jsonify({"ok": False, "error": f"max {MAX_METRICS_REQ} points/request"}), 413
-    now = int(time.time())
-    rows = []
-    for p in pts:
-        try:
-            rows.append((rid, int(p.get("ts") or now), int(p.get("step") or 0),
-                         _clip(p["key"], 128), float(p["value"])))
-        except (KeyError, TypeError, ValueError):
-            continue
-    with LOCK:
-        if not DB.execute("SELECT 1 FROM runs WHERE id=?", (rid,)).fetchone():
-            return jsonify({"ok": False, "error": "unknown run"}), 404
-        if rows:
-            DB.executemany("INSERT INTO run_metrics(run_id,ts,step,key,value) VALUES(?,?,?,?,?)", rows)
-            DB.execute("UPDATE runs SET heartbeat_at=? WHERE id=?", (now, rid))
-            DB.commit()
-    return jsonify({"ok": True, "logged": len(rows)})
-
-@app.route("/api/runs/<rid>/finish", methods=["POST"])
-@require_api_key
-def api_runs_finish(rid):
-    body = request.get_json(silent=True) or {}
-    status = body.get("status") if body.get("status") in RUN_STATUS else "finished"
-    if status == "running":
-        status = "finished"
-    ended = int(body.get("ended_at") or time.time())
-    with LOCK:
-        cur = DB.execute("UPDATE runs SET status=?, ended_at=?, heartbeat_at=? WHERE id=?",
-                         (status, ended, ended, rid))
-        DB.commit()
-    return (jsonify({"ok": True, "id": rid, "status": status}) if cur.rowcount
-            else (jsonify({"ok": False, "error": "unknown run"}), 404))
-
-@app.route("/api/runs/<rid>", methods=["DELETE"])
-def api_runs_delete(rid):
-    """Remove a run and its logged metrics. A same-origin browser management action
-    (like deleting a host or an API key), so it's open on the LAN rather than
-    key-gated — the key gates *ingest* (forgery from notebooks), not housekeeping."""
-    with LOCK:
-        cur = DB.execute("DELETE FROM runs WHERE id=?", (rid,))
-        DB.execute("DELETE FROM run_metrics WHERE run_id=?", (rid,))
-        DB.commit()
-    return (jsonify({"ok": True}) if cur.rowcount
-            else (jsonify({"ok": False, "error": "unknown run"}), 404))
-
-@app.route("/api/runs")
-def api_runs_list():
-    ctx = _cost_ctx()
-    rng = request.args.get("range", "7d"); span = RANGES.get(rng, 604800)
-    status = request.args.get("status")
-    key_filter = request.args.get("key")
-    now = int(time.time()); since = 0 if span is None else now - span
-    q = ("SELECT id,name,source,status,started_at,ended_at,host,params,tags,notes,key_id "
-         "FROM runs WHERE (ended_at IS NULL OR ended_at>=?) AND started_at<=? ")
-    args = [since, now]
-    if status in RUN_STATUS:
-        q += "AND status=? "; args.append(status)
-    if key_filter:
-        q += "AND key_id=? "; args.append(key_filter)
-    q += "ORDER BY started_at DESC LIMIT 500"
-    out = []
-    with LOCK:
-        cur = DB.cursor()
-        key_names = dict(cur.execute("SELECT id, name FROM api_keys").fetchall())
-        for (rid, name, source, st, started, ended, host, params, tags, notes, key_id) in cur.execute(q, args).fetchall():
-            e_kwh, cost, avg_w, peak_u = _run_cost_window(cur, started, ended, ctx)
-            kv = {}
-            # latest value per key = the last-logged row (max rowid), robust even when
-            # several points share a timestamp.
-            for k, v in cur.execute(
-                    "SELECT key, value FROM run_metrics WHERE run_id=? AND rowid IN "
-                    "(SELECT MAX(rowid) FROM run_metrics WHERE run_id=? GROUP BY key)", (rid, rid)):
-                kv[k] = v
-            out.append({"id": rid, "name": name, "source": source, "status": st,
-                        "started_at": started, "ended_at": ended, "duration": (ended or now) - started,
-                        "host": host, "params": _safe_json(params), "tags": _safe_json(tags), "notes": notes,
-                        "key_id": key_id, "key_name": key_names.get(key_id),
-                        "metrics_latest": kv, "energy_kwh": e_kwh, "cost": cost, "avg_w": avg_w, "peak_util": peak_u})
-    return jsonify({"range": rng, "currency": ctx["currency"], "tariff_mode": ctx["mode"], "runs": out})
-
-@app.route("/api/runs/<rid>")
-def api_runs_get(rid):
-    ctx = _cost_ctx()
-    now = int(time.time())
-    with LOCK:
-        cur = DB.cursor()
-        r = cur.execute("SELECT id,name,source,status,started_at,ended_at,host,params,tags,notes "
-                        "FROM runs WHERE id=?", (rid,)).fetchone()
-        if not r:
-            return jsonify({"error": "unknown run"}), 404
-        (rid, name, source, st, started, ended, host, params, tags, notes) = r
-        end = ended or now
-        metrics = {}
-        for k, ts, step, v in cur.execute(
-                "SELECT key,ts,step,value FROM run_metrics WHERE run_id=? ORDER BY key,ts,step", (rid,)):
-            d = metrics.setdefault(k, {"steps": [], "ts": [], "values": []})
-            d["steps"].append(step); d["ts"].append(ts); d["values"].append(v)
-        bk = max(INTERVAL, round(max(1, end - started) / MAX_POINTS))
-        labels, power_w, util_pct = [], [], []
-        for b, ap, au in cur.execute("SELECT (ts/?)*? b, AVG(power), AVG(util) FROM samples "
-                                     "WHERE ts>=? AND ts<=? GROUP BY b ORDER BY b", (bk, bk, started, end)):
-            labels.append(int(b)); power_w.append(round(ap or 0)); util_pct.append(round(au or 0))
-        e_kwh, cost, avg_w, peak_u = _run_cost_window(cur, started, end, ctx)
-    return jsonify({"id": rid, "name": name, "source": source, "status": st,
-                    "started_at": started, "ended_at": ended, "duration": end - started, "host": host,
-                    "params": _safe_json(params), "tags": _safe_json(tags), "notes": notes, "metrics": metrics,
-                    "resource": {"labels": labels, "power_w": power_w, "util_pct": util_pct, "bucket_sec": bk},
-                    "energy_kwh": e_kwh, "cost": cost, "avg_w": avg_w, "peak_util": peak_u,
-                    "currency": ctx["currency"], "tariff_mode": ctx["mode"]})
-
-# ── MLflow sync (pull) — mirror MLflow runs in as source='mlflow', pure REST ───
 _MLF_STATUS = {"RUNNING": "running", "FINISHED": "finished", "FAILED": "failed",
                "KILLED": "killed", "SCHEDULED": "running"}
 
@@ -6083,23 +5349,7 @@ def sync_mlflow():
             break
     return synced
 
-@app.route("/api/integration/mlflow/sync", methods=["GET", "POST"])
-def api_mlflow_sync():
-    """GET -> reachability probe (green/red). POST -> sync now."""
-    if not (get_settings().get("mlflow_uri") or "").strip():
-        return jsonify({"ok": False, "error": "no MLflow URI configured"}), 400
-    if request.method == "POST":
-        try:
-            return jsonify({"ok": True, "synced": sync_mlflow()})
-        except Exception as e:
-            return jsonify({"ok": False, "error": str(e)[:200]}), 502
-    try:
-        _mlf("POST", "/api/2.0/mlflow/experiments/search", {"max_results": 1})
-        return jsonify({"ok": True, "reachable": True})
-    except Exception as e:
-        return jsonify({"ok": False, "reachable": False, "error": str(e)[:200]}), 502
 
-# ── Container logs over SSE (issue #28) ───────────────────────────────────────
 _CT_NAME_RE = re.compile(r"^[a-zA-Z0-9][a-zA-Z0-9_.-]{0,127}$")
 
 def _docker_log_stream(name, tail, follow):
@@ -6159,157 +5409,9 @@ def _docker_log_stream(name, tail, follow):
         try: c.close()
         except Exception: pass
 
-@app.route("/api/containers/<name>/logs")
-def api_container_logs(name):
-    """Last `tail` log lines for a container; with follow=1, streams new lines as
-    SSE. Read-only — `docker logs` needs no extra socket permissions."""
-    if not _CT_NAME_RE.match(name or ""):
-        return jsonify({"error": "invalid container name"}), 400
-    try:
-        tail = max(1, min(2000, int(request.args.get("tail", 200))))
-    except (TypeError, ValueError):
-        tail = 200
-    follow = request.args.get("follow") == "1"
-    return Response(_docker_log_stream(name, tail, follow),
-                    mimetype="text/event-stream",
-                    headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no",
-                             "Connection": "keep-alive"})
-
-@app.route("/api/network")
-def api_network():
-    """Host NIC throughput + per-container top talkers over a range (#30). Rates
-    are derived from the cumulative byte counters in net_samples, so a missed
-    sample or a counter reset (reboot) never invents a spike."""
-    rng = request.args.get("range", "1h")
-    span = RANGES.get(rng, 3600)
-    now = int(time.time())
-    with LOCK:
-        cur = DB.cursor()
-        since = (cur.execute("SELECT MIN(ts) FROM net_samples").fetchone()[0] or now) if span is None else now - span
-        rows = cur.execute("SELECT ts,iface,bytes_in,bytes_out FROM net_samples WHERE ts>=? ORDER BY iface,ts",
-                           (since,)).fetchall()
-    bk = max(INTERVAL, round(max(1, now - since) / MAX_POINTS))
-    series = {}
-    for ts, iface, bi, bo in rows:
-        series.setdefault(iface, []).append((ts, bi, bo))
-
-    def rate_buckets(samples):
-        """Consecutive cumulative samples → {bucket: [sum_rate, count]} for in/out."""
-        ain, aout = {}, {}
-        for (t0, i0, o0), (t1, i1, o1) in zip(samples, samples[1:]):
-            dt = t1 - t0
-            di, do = i1 - i0, o1 - o0
-            if dt <= 0 or di < 0 or do < 0:        # gap or counter reset → skip
-                continue
-            b = (t1 // bk) * bk
-            s = ain.setdefault(b, [0, 0]);  s[0] += di / dt; s[1] += 1
-            s = aout.setdefault(b, [0, 0]); s[0] += do / dt; s[1] += 1
-        return ain, aout
-
-    host_ifaces = sorted(i for i in series if not i.startswith("@") and not _HOST_NIC_SKIP.match(i))
-    labels = sorted({(t // bk) * bk for i in host_ifaces for t, _, _ in series[i]})
-    ifaces_out = []
-    for iface in host_ifaces:
-        ain, aout = rate_buckets(series[iface])
-        ins  = [round(ain[b][0] / ain[b][1]) if ain.get(b, [0, 0])[1] else 0 for b in labels]
-        outs = [round(aout[b][0] / aout[b][1]) if aout.get(b, [0, 0])[1] else 0 for b in labels]
-        if any(ins) or any(outs):
-            ifaces_out.append({"iface": iface, "in": ins, "out": outs})
-
-    talkers = []
-    for iface, samples in series.items():
-        if not iface.startswith("@") or len(samples) < 2:
-            continue
-        di = max(0, samples[-1][1] - samples[0][1])
-        do = max(0, samples[-1][2] - samples[0][2])
-        if di + do > 0:
-            talkers.append({"name": iface[1:], "bytes_in": di, "bytes_out": do, "total": di + do})
-    talkers.sort(key=lambda x: -x["total"])
-
-    cur_in  = sum(s["in"][-1] for s in ifaces_out) if labels else 0
-    cur_out = sum(s["out"][-1] for s in ifaces_out) if labels else 0
-    return jsonify({"range": rng, "bucket_sec": bk, "labels": labels,
-                    "ifaces": ifaces_out, "talkers": talkers[:10],
-                    "current": {"in": cur_in, "out": cur_out}})
-
-@app.route("/healthz")
-def healthz():
-    """Cheap liveness probe for Docker's HEALTHCHECK and any uptime monitor.
-    No DB, no locks — just a 200 with the running version so the answer is
-    instant and never gets blocked behind a slow collector pass."""
-    return jsonify({"status": "ok", "version": VERSION}), 200
-
-@app.route("/api/changelog")
-def api_changelog():
-    """Serve the bundled CHANGELOG.md, sliced to a version range, so the dashboard's
-    one-time 'what's new' modal can show exactly what shipped — straight from the
-    image, no GitHub round-trip (works fully offline). Read-only.
-      ?to=<ver>     newest version to include (default: the running VERSION)
-      ?since=<ver>  exclusive lower bound — return every section newer than it,
-                    up to `to` (the multi-version roll-up). Omit for just `to`."""
-    to_v = request.args.get("to") or VERSION
-    since_v = request.args.get("since")
-    to_t = _parse_semver(to_v)
-    since_t = _parse_semver(since_v) if since_v else None
-    try:
-        path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "CHANGELOG.md")
-        with open(path, "r", encoding="utf-8") as f:
-            text = f.read()
-    except Exception:
-        return jsonify({"current": VERSION, "sections": [], "markdown": ""})
-    # Split on "## [x.y.z](url) — date" headers (Keep-a-Changelog style).
-    hdr = re.compile(r"^##\s*\[([^\]]+)\]\(([^)]*)\)\s*[—\-–]?\s*(.*)$")
-    sections, cur = [], None
-    for line in text.splitlines():
-        m = hdr.match(line)
-        if m:
-            if cur:
-                sections.append(cur)
-            cur = {"version": m.group(1).strip(), "url": m.group(2).strip(),
-                   "date": m.group(3).strip(), "lines": [line]}
-        elif cur is not None:
-            cur["lines"].append(line)
-    if cur:
-        sections.append(cur)
-    picked = []
-    for s in sections:                       # file is newest-first
-        sv = _parse_semver(s["version"])
-        if sv > to_t:
-            continue
-        if since_t is not None:
-            if sv > since_t:
-                picked.append(s)
-        else:
-            picked.append(s)                 # no lower bound → just the newest <= to
-            break
-    md = "\n".join("\n".join(s["lines"]).rstrip() for s in picked)
-    return jsonify({"current": VERSION, "to": to_v, "since": since_v,
-                    "sections": [{"version": s["version"], "date": s["date"], "url": s["url"]}
-                                 for s in picked],
-                    "markdown": md})
-
-@app.route("/favicon.ico")
-def favicon():
-    """Default-favicon URL — browsers ask for /favicon.ico even when an explicit
-    <link rel="icon"> points elsewhere (during early page load, or for tabs that
-    open without rendering HTML). Serve the SVG we ship in static/."""
-    return app.send_static_file("favicon.svg")
 
 _LOCALES_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), "locales")
 
-@app.route("/locales/<path:fn>")
-def locales(fn):
-    """Serve UI translation files (i18n, #148). The dashboard fetches
-    /locales/<code>.json for any non-English locale; English is inlined, so it
-    needs no fetch. send_from_directory guards against path traversal."""
-    if not fn.endswith(".json"):
-        return ("Not found", 404)
-    try:
-        resp = send_from_directory(_LOCALES_DIR, fn)
-    except Exception:
-        return ("Not found", 404)
-    resp.headers["Cache-Control"] = "no-cache"
-    return resp
 
 def _mcp_enabled():
     return os.environ.get("ENABLE_MCP", "1").strip().lower() not in ("0", "false", "no")
@@ -6357,118 +5459,6 @@ def build_mcp_status():
         out["last_activity_age_s"] = age_s
     return out
 
-@app.route("/api/mcp-status")
-def api_mcp_status():
-    return jsonify(build_mcp_status())
-
-@app.route("/api/health")
-def api_health():
-    """Current state of the status monitors (Docker + systemd) plus a light GPU/host
-    snapshot. Cheap and DB-free, so the dashboard can poll it often."""
-    gpu_avail = LATEST.get("gpu_avail")
-    now = {"gpu": {"util": LATEST["util"], "mem_used": LATEST["mem_used"],
-                   "mem_total": (LATEST["mem_total"] or 24576) if gpu_avail else 0,
-                   "power": LATEST["power"], "temp": LATEST["temp"],
-                   "available": bool(gpu_avail),
-                   "gpus": LATEST.get("gpus") or [],    # per-card detail (issue #95)
-                   "extra": LATEST.get("gpu_extra") or {}},  # mem-bw/clocks/throttle (telemetry)
-           "host": enrich_os_upgrade(LATEST["host"])}
-    docker  = HEALTH["docker"]  or {"available": False, "reason": "warming up…",
-                                    "containers": [], "summary": {"total": 0, "running": 0, "problems": 0}}
-    systemd = HEALTH["systemd"] or {"available": False, "reason": "warming up…",
-                                    "services": [], "summary": {}}
-    update  = dict(HEALTH["update"] or {"available": False, "current": VERSION})
-    # Let the frontend decide whether to show the one-click "Update now" button.
-    # Set here (not baked into the cached collect_update payload) so toggling the
-    # env flag takes effect on restart without waiting for the update cache.
-    update["self_update_enabled"] = ALLOW_SELF_UPDATE
-    disk_io = dict(HEALTH.get("disk_io") or {"available": False, "warming_up": True,
-                                              "summary": {"total_read_mb_s": 0.0, "total_write_mb_s": 0.0},
-                                              "items": []})
-    # Per-process I/O attribution (Top writer/reader) — attach ONLY to this authed
-    # payload. Carries process comm (never cmdline/argv) and NEVER appears on the
-    # public status surface (build_public_status doesn't read processes/disk_io).
-    _pio = (HEALTH.get("processes") or {}).get("io")
-    if _pio and _pio.get("available"):
-        disk_io["attribution"] = _pio
-    if _diskio_anom_latest:
-        disk_io["anomalies"] = dict(_diskio_anom_latest)
-    return jsonify({"version": VERSION, "updated": HEALTH["at"], "now": now,
-                    "docker": docker, "systemd": systemd, "update": update,
-                    "processes": HEALTH["processes"],
-                    "disk_io": disk_io,
-                    "os_updates": os_updates_summary(),
-                    "diagnostics": local_diagnostics(),
-                    "mcp": {"enabled": _mcp_enabled(), "port": _mcp_port()},
-                    "overview": build_overview(now, docker, systemd)})
-
-@app.route("/metrics")
-def metrics():
-    """Prometheus text-format scrape endpoint.
-
-    Reads exclusively from the in-memory snapshots (LATEST / HEALTH) that the
-    background collector keeps fresh.  No new I/O is triggered on each scrape,
-    so double-sampling is impossible.
-    """
-    if not _PROM_OK:
-        return Response("# prometheus_client not installed\n", mimetype="text/plain", status=503)
-
-    # Clear all multi-label gauges before re-populating so stale series vanish.
-    for key in ("gpu_vram_used", "gpu_vram_total", "gpu_util", "gpu_temp", "gpu_power",
-                "host_disk_used", "model_vram", "container_state", "systemd_unit",
-                "models_installed"):
-        _G[key].clear()
-
-    # ── GPU ──────────────────────────────────────────────────────────────────
-    gpu_label = "gpu0"
-    _G["gpu_vram_used"].labels(gpu=gpu_label).set(LATEST.get("mem_used", 0))
-    _G["gpu_vram_total"].labels(gpu=gpu_label).set(LATEST.get("mem_total", 0))
-    _G["gpu_util"].labels(gpu=gpu_label).set(LATEST.get("util", 0))
-    _G["gpu_temp"].labels(gpu=gpu_label).set(LATEST.get("temp", 0))
-    _G["gpu_power"].labels(gpu=gpu_label).set(LATEST.get("power", 0))
-
-    # ── Host ─────────────────────────────────────────────────────────────────
-    host = LATEST.get("host") or {}
-    _G["host_cpu"].set(host.get("cpu", 0))
-    ram_total = host.get("ram_total") or 1
-    ram_used  = host.get("ram_used", 0)
-    _G["host_mem_used"].set(round(100 * ram_used / ram_total, 1))
-    for disk in (host.get("disks") or []):
-        _G["host_disk_used"].labels(mountpoint=disk["mount"]).set(disk.get("pct", 0))
-
-    # ── Model VRAM ───────────────────────────────────────────────────────────
-    for entry in (LATEST.get("models") or []):
-        vram = entry.get("vram")
-        if vram is not None:
-            _G["model_vram"].labels(server=entry.get("service", "?"),
-                                    model=entry.get("model", "?")).set(vram)
-
-    # ── Installed-models registry, by provider (#219) — no new I/O: counts the
-    # already-sampled model_catalog, same source the /api/models endpoint merges.
-    provider_counts = {}
-    for entry in (LATEST.get("model_catalog") or []):
-        p = entry.get("provider") or "unknown"
-        provider_counts[p] = provider_counts.get(p, 0) + 1
-    for provider, count in provider_counts.items():
-        _G["models_installed"].labels(provider=provider).set(count)
-
-    # ── Docker containers ────────────────────────────────────────────────────
-    docker = HEALTH.get("docker") or {}
-    for ct in (docker.get("containers") or []):
-        name  = ct.get("name", "?")
-        state = ct.get("state", "unknown")
-        _G["container_state"].labels(name=name, state=state).set(1)
-
-    # ── Systemd units ────────────────────────────────────────────────────────
-    systemd = HEALTH.get("systemd") or {}
-    for svc in (systemd.get("services") or []):
-        unit   = svc.get("name", "?")
-        active = svc.get("active", "unknown")
-        _G["systemd_unit"].labels(unit=unit, state=active).set(
-            1 if active == "active" else 0)
-
-    return Response(generate_latest(), mimetype=CONTENT_TYPE_LATEST)
-
 
 def _public_settings():
     """Same as get_settings(), but redacts secrets and reports their presence."""
@@ -6478,69 +5468,7 @@ def _public_settings():
         out[k + "_set"] = bool(s.get(k))
     return out
 
-@app.route("/api/hub/pubkey")
-def api_hub_pubkey():
-    return jsonify({"pubkey": get_hub_pubkey()})
 
-@app.route("/api/hosts", methods=["GET", "POST"])
-def api_hosts():
-    if request.method == "POST":
-        body = request.get_json(silent=True) or {}
-        host, err = add_host((body.get("name") or "").strip(),
-                             (body.get("ssh_target") or "").strip(),
-                             (body.get("tags") or "").strip())
-        if err:
-            return jsonify({"ok": False, "error": err}), 400
-        return jsonify({"ok": True, "host": host}), 201
-    return jsonify({"hosts": list_hosts()})
-
-@app.route("/api/hosts/<name>", methods=["DELETE", "PATCH"])
-def api_hosts_one(name):
-    if request.method == "DELETE":
-        ok = delete_host(name)
-        return jsonify({"ok": ok}), (200 if ok else 404)
-    body = request.get_json(silent=True) or {}
-    host, err = update_host(
-        name,
-        ssh_target=(body.get("ssh_target").strip() if isinstance(body.get("ssh_target"), str) else None),
-        tags=(body.get("tags").strip() if isinstance(body.get("tags"), str) else None),
-    )
-    if err:
-        return jsonify({"ok": False, "error": err}), 400 if "look like" in err or "Nothing" in err else 404
-    return jsonify({"ok": True, "host": host})
-
-@app.route("/api/lan/scan")
-def api_lan_scan():
-    return jsonify(discover_lan())
-
-@app.route("/api/host_data/<name>")
-def api_host_data(name):
-    """Per-host metric snapshot. `local` is the hub itself; remotes are served
-    from the in-memory cache populated by host_poller()."""
-    if name == "local":
-        return jsonify({"name": "local", "host": enrich_os_upgrade(_local_now_snapshot()),
-                        "at": int(time.time()), "online": True})
-    with HOST_DATA_LOCK:
-        entry = HOST_DATA.get(name)
-    if not entry or "data" not in entry:
-        # No successful poll yet (or never registered) — still respond 200 so
-        # the UI can render a "waiting" state instead of erroring.
-        return jsonify({"name": name, "online": False,
-                        "error": (entry or {}).get("error") or "no data yet",
-                        "at": (entry or {}).get("at"),
-                        "host": None})
-    age     = int(time.time()) - int(entry["at"])
-    online  = _host_is_online(entry)
-    return jsonify({"name": name, "host": enrich_os_upgrade(entry["data"].get("host", {})),
-                    "at": entry["at"], "online": online,
-                    "stale_for": 0 if online else age,
-                    "error": entry.get("error")})
-
-# ── On-demand disk-usage scan (WizTree-style folder treemap) ──────────────────
-# `du --max-depth=1` of a host path (read through the read-only HOST_ROOT mount),
-# run in a background thread so a slow scan of a big disk never blocks a request.
-# The UI polls until state=="done". Results cached per path; one filesystem only
-# (--one-file-system) so scanning "/" doesn't wander into other mounted disks.
 _DISK_SCAN, _DISK_SCAN_LOCK = {}, threading.Lock()
 _DISK_SCAN_TTL = 900   # reuse a completed scan for 15 min
 
@@ -6614,361 +5542,6 @@ def _disk_scan_worker(path, real):
         with _DISK_SCAN_LOCK:
             _DISK_SCAN[path] = {"state": "error", "at": int(time.time()), "error": str(e)[:200]}
 
-@app.route("/api/disk_scan")
-def api_disk_scan():
-    path = os.path.normpath(request.args.get("path") or "/")
-    rescan = request.args.get("rescan") == "1"
-    real = _safe_host_dir(path)
-    if not real:
-        return jsonify({"path": path, "state": "error", "error": f"not a readable directory: {path}"})
-    with _DISK_SCAN_LOCK:
-        ent = _DISK_SCAN.get(path)
-        if ent and not rescan:
-            if ent["state"] == "scanning":
-                return jsonify({"path": path, "state": "scanning"})
-            if ent["state"] == "done" and time.time() - ent["at"] < _DISK_SCAN_TTL:
-                return jsonify({"path": path, **ent})
-            if ent["state"] == "error" and time.time() - ent["at"] < 20:
-                return jsonify({"path": path, **ent})
-        _DISK_SCAN[path] = {"state": "scanning", "at": int(time.time())}
-    threading.Thread(target=_disk_scan_worker, args=(path, real), daemon=True).start()
-    return jsonify({"path": path, "state": "scanning"})
-
-@app.route("/api/fleet")
-def api_fleet():
-    """Compact summary KPIs for every host in the fleet. Drives the All-hosts
-    table. Order: local first, then registered hosts in the order they were
-    added."""
-    hosts = list_hosts()
-    rows  = []
-
-    # Local row
-    rows.append({"name": "local", "label": socket.gethostname() + " (this hub)",
-                 "ssh_target": None, "host": enrich_os_upgrade(_local_now_snapshot()),
-                 "at": int(time.time()), "online": True, "is_local": True,
-                 "last_check": {"summary": {"overall": "ok"}}})
-
-    with HOST_DATA_LOCK:
-        for h in hosts:
-            entry = HOST_DATA.get(h["name"]) or {}
-            data  = entry.get("data") or {}
-            at    = entry.get("at")
-            online = _host_is_online(entry)
-            rows.append({
-                "name": h["name"],
-                "label": h["name"],
-                "ssh_target": h["ssh_target"],
-                "host": enrich_os_upgrade(data.get("host")) if data else None,
-                "at": at,
-                "online": online,
-                "is_local": False,
-                "last_check": h.get("last_check"),
-                "error": entry.get("error"),
-            })
-    return jsonify({"hosts": rows, "interval": INTERVAL})
-
-@app.route("/api/hosts/<name>/test", methods=["POST"])
-def api_hosts_test(name):
-    result = probe_host(name)
-    if result is None:
-        return jsonify({"ok": False, "error": "no such host"}), 404
-    return jsonify({"ok": True, "result": result})
-
-@app.route("/api/hosts/<name>/run", methods=["POST"])
-def api_hosts_run(name):
-    """Execute a command on a registered host. Body: {cmd, sudo_password?}.
-    The sudo password is processed in-memory: piped via stdin to `sudo -S`
-    on the remote, NEVER stored in the DB, NEVER written to logs, NEVER
-    in any process's argv. Don't pass arbitrary cmd from untrusted users —
-    the API is reachable to anyone on the LAN who can hit the dashboard."""
-    body = request.get_json(silent=True) or {}
-    cmd = (body.get("cmd") or "").strip()
-    sudo_password = body.get("sudo_password") or None
-    if not cmd:
-        return jsonify({"ok": False, "error": "cmd required"}), 400
-    result = run_on_host(name, cmd, sudo_password=sudo_password)
-    # Drop the password reference ASAP — Python keeps the string object until
-    # GC, but at least we don't hold our own reference past this point.
-    sudo_password = None
-    body = None
-    if result is None:
-        return jsonify({"ok": False, "error": "no such host"}), 404
-    return jsonify(result)
-
-@app.route("/api/backup")
-def api_backup_download():
-    """Stream a consistent SQLite snapshot (VACUUM INTO) of the live database."""
-    if _DB_MAINTENANCE:
-        return jsonify({"ok": False, "error": "Database maintenance in progress."}), 503
-    if not _data_dir_writable():
-        return jsonify({"ok": False, "error": "Cannot write backup — mount a writable /data volume."}), 400
-    tmp_path = None
-    try:
-        with LOCK:
-            fd, tmp_path = tempfile.mkstemp(suffix=".db", prefix=".backup_", dir=_data_dir())
-            os.close(fd)
-            db_backup.vacuum_into(DB, tmp_path)
-    except Exception as e:
-        if tmp_path:
-            try: os.unlink(tmp_path)
-            except OSError: pass
-        return jsonify({"ok": False, "error": "Backup failed: %s" % e}), 500
-
-    @after_this_request
-    def _cleanup_snapshot(resp):
-        try: os.unlink(tmp_path)
-        except OSError: pass
-        return resp
-
-    return send_file(tmp_path, mimetype="application/x-sqlite3", as_attachment=True,
-                     download_name=db_backup.backup_filename())
-
-@app.route("/api/backup/restore", methods=["POST"])
-def api_backup_restore():
-    """Replace the live database with an uploaded backup snapshot."""
-    global _DB_MAINTENANCE
-    if _DB_MAINTENANCE:
-        return jsonify({"ok": False, "error": "Database maintenance already in progress."}), 503
-    if not _data_dir_writable():
-        return jsonify({"ok": False, "error": "Cannot restore — mount a writable /data volume."}), 400
-    upload = request.files.get("backup")
-    if not upload or not upload.filename:
-        return jsonify({"ok": False, "error": "No backup file uploaded."}), 400
-
-    upload_path = None
-    try:
-        fd, upload_path = tempfile.mkstemp(suffix=".db", prefix=".restore_upload_", dir=_data_dir())
-        os.close(fd)
-        upload.save(upload_path)
-        ok, err = db_backup.validate_backup(upload_path)
-        if not ok:
-            return jsonify({"ok": False, "error": err}), 400
-
-        with LOCK:
-            _DB_MAINTENANCE = True
-            try:
-                if os.path.isfile(DB_PATH):
-                    shutil.copy2(DB_PATH, "%s.pre-restore-%d.bak" % (DB_PATH, int(time.time())))
-                try:
-                    DB.close()
-                except Exception:
-                    pass
-                db_backup.remove_wal_sidecars(DB_PATH)
-                os.replace(upload_path, DB_PATH)
-                upload_path = None
-                db_backup.remove_wal_sidecars(DB_PATH)
-                reopen_db()
-            except Exception as e:
-                try:
-                    reopen_db()
-                except Exception:
-                    pass
-                return jsonify({"ok": False, "error": "Restore failed: %s" % e}), 500
-            finally:
-                _DB_MAINTENANCE = False
-
-        return jsonify({"ok": True,
-                        "message": "Backup restored. History and settings have been reloaded."})
-    finally:
-        if upload_path:
-            try: os.unlink(upload_path)
-            except OSError: pass
-
-@app.route("/api/settings", methods=["GET", "POST"])
-def api_settings():
-    if request.method == "POST":
-        body = request.get_json(silent=True) or {}
-        # Empty string clears a setting; missing key leaves it unchanged.
-        # Secrets pass through the "_set: false" sentinel from the UI as a way
-        # to clear without revealing the current value.
-        updates = {k: body[k] for k in body if k in SETTING_DEFAULTS}
-        err = (_validate_url_settings(updates) or _validate_email_settings(updates)
-               or _validate_brief_settings(updates))
-        if err:
-            return jsonify({"ok": False, "error": err}), 400
-        save_settings(updates)
-    return jsonify({"version": VERSION, "settings": _public_settings()})
-
-@app.route("/api/notify/test", methods=["POST"])
-def api_notify_test():
-    """Send a one-shot test alert using the currently saved settings."""
-    s = get_settings()
-    if not (s.get("discord_webhook_url") or s.get("ntfy_topic")
-            or (s.get("telegram_token") and s.get("telegram_chat_id"))
-            or (s.get("email_host") and s.get("email_from") and s.get("email_to"))
-            or s.get("slack_webhook_url")
-            or s.get("webhook_url")):
-        return jsonify({"ok": False, "results": [],
-                        "reason": "No Discord webhook, ntfy topic, Telegram bot, email, Slack webhook, or generic webhook configured."}), 400
-    results = dispatch_alert(s, "info",
-                             "✅ HomeLab Monitor — test alert",
-                             "If you see this, alerts are wired up correctly.")
-    return jsonify({"ok": all(ok for _, ok, _ in results),
-                    "results": [{"channel": c, "ok": ok, "error": err} for c, ok, err in results]})
-
-@app.route("/api/uptime", methods=["GET", "POST"])
-def api_uptime():
-    """List uptime checks with their live state (GET) or create one (POST).
-    GET is always 200 (an empty list before any are added); POST returns a clean
-    400 with a human error on invalid input."""
-    if request.method == "POST":
-        body = request.get_json(silent=True) or {}
-        cid, err = create_uptime_check(body)
-        if err:
-            return jsonify({"ok": False, "error": err}), 400
-        return jsonify({"ok": True, "id": cid}), 201
-    return jsonify(uptime_overview())
-
-@app.route("/api/uptime/<cid>", methods=["PATCH", "DELETE"])
-def api_uptime_one(cid):
-    """Update (full edit or a quick enabled toggle) or delete one check."""
-    if request.method == "DELETE":
-        return (jsonify({"ok": True}) if delete_uptime_check(cid)
-                else (jsonify({"ok": False, "error": "not found"}), 404))
-    body = request.get_json(silent=True) or {}
-    ok, err = update_uptime_check(cid, body)
-    if ok:
-        return jsonify({"ok": True})
-    return jsonify({"ok": False, "error": err}), (404 if err == "not found" else 400)
-
-@app.route("/api/maintenance", methods=["GET", "POST"])
-def api_maintenance():
-    """List maintenance windows (GET) or create one (POST)."""
-    if request.method == "POST":
-        body = request.get_json(silent=True) or {}
-        label = (body.get("label") or "").strip()
-        if not label:
-            return jsonify({"ok": False, "error": "label is required"}), 400
-        start_ts = body.get("start_ts")
-        end_ts   = body.get("end_ts")
-        if not start_ts or not end_ts:
-            return jsonify({"ok": False, "error": "start_ts and end_ts are required"}), 400
-        try:
-            start_ts = int(start_ts)
-            end_ts = int(end_ts)
-        except (TypeError, ValueError):
-            return jsonify({"ok": False, "error": "start_ts and end_ts must be numeric"}), 400
-        if end_ts <= start_ts:
-            return jsonify({"ok": False, "error": "end_ts must be after start_ts"}), 400
-        recurrence = body.get("recurrence") or None
-        if recurrence and recurrence not in ("daily", "weekly"):
-            return jsonify({"ok": False, "error": "recurrence must be daily, weekly, or null"}), 400
-        if recurrence == "daily" and (end_ts - start_ts) > 86400:
-            return jsonify({"ok": False, "error": "daily recurrence window cannot exceed 24h"}), 400
-        if recurrence == "weekly" and (end_ts - start_ts) > 604800:
-            return jsonify({"ok": False, "error": "weekly recurrence window cannot exceed 7d"}), 400
-        wid = create_maintenance_window(
-            label=label,
-            kind=body.get("kind") or "*",
-            pattern=body.get("pattern") or "*",
-            start_ts=start_ts,
-            end_ts=end_ts,
-            recurrence=recurrence,
-            note=body.get("note"),
-        )
-        return jsonify({"ok": True, "id": wid}), 201
-    return jsonify(list_maintenance_windows())
-
-@app.route("/api/maintenance/<wid>", methods=["DELETE"])
-def api_maintenance_one(wid):
-    """Delete a maintenance window."""
-    delete_maintenance_window(wid)
-    return jsonify({"ok": True})
-
-@app.route("/api/update/app", methods=["POST"])
-def api_update_app():
-    """Start the opt-in one-click self-update (detached docker:cli helper).
-    400 = disabled / no update / not a compose deploy; 409 = already running;
-    202 = job started. Gated by ALLOW_SELF_UPDATE — off by default."""
-    body = request.get_json(silent=True) or {}
-    force = bool(body.get("force"))
-    code, payload = start_self_update(force=force)
-    return jsonify(payload), code
-
-@app.route("/api/update/app/status")
-def api_update_app_status():
-    """Read back the self-update progress files from the data dir. Works even
-    right after the restart — it just reads update_state.json + a tail of
-    update.log. No state file yet → idle."""
-    st = _read_update_state()
-    if not st:
-        return jsonify({"state": "idle"})
-    st = dict(st)
-    st["log"] = _tail_lines(_update_log_path(), 200)
-    st["self_update_enabled"] = ALLOW_SELF_UPDATE
-    return jsonify(st)
-
-@app.route("/api/notify/rules", methods=["GET", "POST"])
-def api_notify_rules():
-    if request.method == "POST":
-        body = request.get_json(silent=True) or {}
-        action = body.get("action", "add")
-        if action == "add":
-            with LOCK:
-                DB.execute("INSERT INTO notification_rules (match_kind, match_pattern, channel, min_level, enabled) VALUES (?, ?, ?, ?, ?)",
-                           (body.get("match_kind", "container"), body.get("match_pattern", "*"),
-                            body.get("channel", "all"), body.get("min_level", "warning"),
-                            1 if body.get("enabled", True) else 0))
-                DB.commit()
-        elif action == "update":
-            rule_id = body.get("id")
-            if not rule_id:
-                return jsonify({"ok": False, "error": "id required"}), 400
-            with LOCK:
-                DB.execute("UPDATE notification_rules SET match_kind=?, match_pattern=?, channel=?, min_level=?, enabled=? WHERE id=?",
-                           (body.get("match_kind"), body.get("match_pattern"),
-                            body.get("channel"), body.get("min_level"),
-                            1 if body.get("enabled", True) else 0, rule_id))
-                DB.commit()
-        elif action == "delete":
-            rule_id = body.get("id")
-            if not rule_id:
-                return jsonify({"ok": False, "error": "id required"}), 400
-            with LOCK:
-                DB.execute("DELETE FROM notification_rules WHERE id=?", (rule_id,))
-                DB.commit()
-        else:
-            return jsonify({"ok": False, "error": f"unknown action: {action}"}), 400
-        return jsonify({"ok": True, "rules": get_notification_rules()})
-    return jsonify({"rules": get_notification_rules()})
-
-@app.route("/api/notify/rules/test", methods=["POST"])
-def api_notify_rules_test():
-    """Test a notification rule by sending a sample alert that would match it."""
-    body = request.get_json(silent=True) or {}
-    s = get_settings()
-    if not (s.get("discord_webhook_url") or s.get("ntfy_topic")
-            or (s.get("telegram_token") and s.get("telegram_chat_id"))):
-        return jsonify({"ok": False, "error": "No notification channels configured."}), 400
-    test_rule = {
-        "match_kind": body.get("match_kind", "container"),
-        "match_pattern": body.get("match_pattern", "*"),
-        "channel": body.get("channel", "all"),
-        "min_level": body.get("min_level", "warning"),
-        "enabled": True,
-    }
-    test_key = f"{test_rule['match_kind']}:test-rule"
-    channels = _apply_rules(test_key, body.get("level", "warning"), [test_rule])
-    if channels is None:
-        return jsonify({"ok": False, "error": "Rule would not match — check kind and pattern."}), 400
-    _dispatch_to_channels(s, body.get("level", "warning"),
-                          "🔔 HomeLab Monitor — rule test",
-                          f"Test of rule: {test_rule['match_kind']} / {test_rule['match_pattern']} → {test_rule['channel']} @ {test_rule['min_level']}",
-                          channels)
-    return jsonify({"ok": True, "channels": list(channels)})
-
-
-@app.route("/")
-def index():
-    return app.send_static_file("dashboard.html")
-
-# ── Daily brief (#170) ───────────────────────────────────────────────────────
-# An opt-in, once-a-day HTML health digest assembled from data we already collect
-# (overview + fleet + uptime checks + recent events + GPU/cost). Email gets the
-# full self-contained, inline-styled HTML (survives Gmail/Outlook with no external
-# assets); chat channels get a compact text summary with the things that need
-# attention. Pure stdlib. Each section degrades gracefully when its data is absent.
-# (_BRIEF_CHANNELS is defined up by the settings validators that consume it.)
 
 _BRIEF_PALETTE = {
     "dark":  {"bg": "#0d1117", "card": "#161b22", "bd": "#30363d", "sub": "#21262d",
@@ -7373,34 +5946,7 @@ def _brief_run_once(now=None):
 # Phase 3.2: moved to backend/collectors/ — re-exported for backward compat
 from backend.collectors import brief_worker
 
-@app.route("/api/brief/preview")
-def api_brief_preview():
-    """Render the current brief as HTML (the Preview button opens this in a tab)."""
-    theme = request.args.get("theme") or get_settings().get("brief_theme", "dark")
-    html, _, _, _ = render_brief("light" if theme == "light" else "dark")
-    return Response(html, mimetype="text/html; charset=utf-8")
 
-@app.route("/api/brief/test", methods=["POST"])
-def api_brief_test():
-    """Render + deliver a brief now to the chosen channel (Send-test button)."""
-    s = get_settings()
-    body = request.get_json(silent=True) or {}
-    ch = (body.get("channel") or s.get("brief_channel") or "").strip()
-    if ch not in _BRIEF_CHANNELS:
-        return jsonify({"ok": False, "error": "Pick a channel for the daily brief first."}), 400
-    if not _brief_channel_ready(s, ch):
-        return jsonify({"ok": False, "error": f"The {ch} channel isn't configured yet — fill it in above under Alerts."}), 400
-    try:
-        send_brief(s, ch)
-    except Exception as e:
-        return jsonify({"ok": False, "error": str(e)}), 502
-    return jsonify({"ok": True, "channel": ch})
-
-# Under pytest, "pytest" is already in sys.modules by the time any test file's
-# `import app` runs — skip the live background workers there. They mutate global
-# state (LATEST/HEALTH/DB) on their own clock with no test-mode awareness, so a
-# long-enough suite could otherwise race a real collector tick against a test's
-# own explicit state setup. Never gated in production (no pytest import there).
 if "pytest" not in sys.modules:
     threading.Thread(target=collector, daemon=True).start()
     threading.Thread(target=host_poller, daemon=True).start()
@@ -7583,50 +6129,6 @@ def _public_overall_status(cards, monitors):
         return "crit"
     return "ok" if all(c.get("status") == "ok" for c in cards) else "warn"
 
-@app.route("/api/public-status")
-def api_public_status():
-    if not _os.environ.get("PUBLIC_STATUS"):
-        abort(404)
-    cfg = get_settings()
-    gpu_avail = LATEST.get("gpu_avail")
-    now = {"gpu": {"util": LATEST.get("util"), "mem_used": LATEST.get("mem_used"),
-                   "mem_total": (LATEST.get("mem_total") or 24576) if gpu_avail else 0,
-                   "power": LATEST.get("power"), "temp": LATEST.get("temp"),
-                   "available": bool(gpu_avail),
-                   "gpus": LATEST.get("gpus") or [],
-                   "extra": LATEST.get("gpu_extra") or {}},
-           "host": LATEST.get("host") or {}}
-    docker  = HEALTH.get("docker")  or {"available": False, "reason": "warming up", "containers": [], "summary": {"total": 0, "running": 0, "problems": 0}}
-    systemd = HEALTH.get("systemd") or {"available": False, "reason": "warming up", "services": [], "summary": {}}
-    cards = build_overview(now, docker, systemd)
-    safe_cards = [{k: v for k, v in c.items() if k in ("key", "label", "status", "metric", "detail")} for c in cards]
-    _now = int(time.time())
-    monitors = _public_monitors(_now)
-    return jsonify({
-        "lab_name": cfg.get("lab_name", "My HomeLab"),
-        "lab_emoji": cfg.get("lab_emoji", "🏠"),
-        "overview": safe_cards,
-        "monitors": monitors,
-        "monitors_summary": _public_monitors_summary(monitors),
-        "incidents": _public_incident_feed(monitors, _now),
-        "status": _public_overall_status(safe_cards, monitors),
-    })
-
-@app.route("/api/public-status/<cid>")
-def api_public_status_one(cid):
-    if not _os.environ.get("PUBLIC_STATUS"):
-        abort(404)
-    detail = _public_status_detail(cid, int(time.time()))
-    if detail is None:
-        abort(404)
-    return jsonify(detail)
-
-@app.route("/public")
-@app.route("/public/<cid>")
-def public_status(cid=None):
-    if not _os.environ.get("PUBLIC_STATUS"):
-        abort(404)
-    return send_from_directory("static", "public.html")
 
 if __name__ == "__main__":
     print(
